pyproject.toml
[project]
name = "embedder"
version = "0.0.1"
description = "Web service of embedder"
authors = [
    {name = "Evgeny",email = "null"}
]
package-mode = false
requires-python = ">=3.12"
dependencies = [
    "fastapi (>=0.136.0,<0.137.0)",
    "uvicorn (>=0.44.0,<0.45.0)",
    "fastembed (>=0.8.0,<0.9.0)",
    "pydantic-settings (>=2.13.1,<3.0.0)",
    "python-dotenv (>=1.2.2,<2.0.0)",
    "loguru (>=0.7.3,<0.8.0)",
    "ruff (>=0.15.11,<0.16.0)",
    "aiofiles (>=25.1.0,<26.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
